requirements.txt
Adafruit-Blinka==8.15.2
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.8
adafruit-circuitpython-pixelbuf==2.0.0
adafruit-circuitpython-requests==1.13.0
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.41.0
Adafruit-PureIO==1.1.10
appdirs==1.4.4
asgiref==3.2.10
Babel==2.8.0
boto3==1.18.28
botocore==1.21.28
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
configparser==5.3.0
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.45.0
django-bootstrap-form==3.4
django-classified==0.9
django-cors-headers==3.8.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-static-fontawesome==5.12.1.4
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
mysql-connector-python==8.0.31
numpy==1.19.2
oauthlib==3.1.1
paho-mqtt==1.6.1
Pillow==7.1.2
protobuf==3.20.1
psycopg2==2.9.1
pycparser==2.20
pyftdi==0.54.0
PyJWT==2.1.0
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
pyusb==1.2.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
scipy==1.5.2
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
ttkthemes==3.2.2
typing_extensions==4.5.0
Unidecode==1.1.1
urllib3==1.26.6
vboxapi==1.0
virtualenv==20.0.30
whitenoise==5.2.0
